refactor options pkg
